+++ libs/adapter/index.ts
@@ -1,3 +1,5 @@
+export * from './cheerio/cheerio-service';
+
 export * from './entity-repository/in-memory.entity-repository';
 export * from './entity-repository/mongo-db.entity-repository';
 
@@ -9,7 +11,7 @@ export * from './feature-service/in-memory.feature-service';
 export * from './file-store/mongo-db.file-store';
 export * from './file-store/network.exception';
 export * from './file-store/s3.file-store';
-export * from './file-store/uri.file-store';
+export * from './file-store/url.file-store';
 
 export * from './logging-service/mongo-db.logging-service';
 export * from './mongo-db/mongo-db.helper';