+++ libs/core/lib/cheerio-service.ts
@@ -1,0 +1,5 @@
+import { CheerioAPI } from 'cheerio';
+
+export interface CheerioService {
+  getByUrl(url: string): Promise<CheerioAPI>;
+}