+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.66
+
+- feat: add cheerio-service type and adapter
+- refactor: rename uri.file-store to url.file-store
+- refactor: export uri.file-store instance
+
 ## 0.0.65
 
 - refactor: metamask plugin and guard